[MCHANGES-203] Use the POMs issueManagement.system when selecting the system to use
The way this is implemented may unfortunately mean that some users will need to reconfigure the plugin if they upgrade. This is because we still have to carry along some backward compatibility luggage. I have added instructions for this on the usage page. If you look at the changes made to the ITs in this commit, you'll see what kind of changes I'm talking about.

git-svn-id: http://svn.apache.org/repos/asf/maven/plugins/trunk/maven-changes-plugin@1054461 13f79535-47bb-0310-9956-ffa450edef68